analytics: Extract analytics/views/support.py.
