+++ analytics/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import include
 from django.urls import path
 
-from analytics.views.legacy import get_activity, get_realm_activity, get_user_activity, support
+from analytics.views.legacy import get_activity, get_realm_activity, get_user_activity
 from analytics.views.stats import (
     get_chart_data,
     get_chart_data_for_installation,
@@ -14,6 +14,7 @@ from analytics.views.stats import (
     stats_for_remote_installation,
     stats_for_remote_realm,
 )
+from analytics.views.support import support
 from zerver.lib.rest import rest_path
 
 i18n_urlpatterns = [
